ImageAnalyzer/bin/requirements.txt
ImageHash~=4.2.0
future~=0.18.2
Pillow~=8.0.1
selenium~=3.141.0
opencv-python~=4.5.1.48
imageio~=2.9.0
matplotlib~=3.3.4
numpy~=1.20.1